# BackEnd/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "马术障碍赛路线设计器 API",
        "version": "1.0.0",
        "description": "路线设计、用户、会员、支付、AI 生成、反馈和协作接口"
    },
    "paths": {
...
        "/user/obstacles/{id}/toggle-share/": {
            "post": {
                "operationId": "user_obstacles_toggle_share_create",
                "description": "切换障碍物的共享状态",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "标识此 自定义障碍物 的 唯一整数值。",
                        "required": true
                    }
                ],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomObstacleRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomObstacleRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomObstacleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "CookieJWTAuth": []
                    },
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomObstacle"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/user/obstacles/count/": {
            "get": {
                "operationId": "user_obstacles_count_retrieve",
                "description": "获取用户自定义障碍物数量和限制",
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "CookieJWTAuth": []
                    },
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomObstacle"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/user/obstacles/shared/": {
            "get": {
                "operationId": "user_obstacles_shared_retrieve",
                "description": "获取其他用户共享的障碍物\n支持分页、搜索和排序",
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "CookieJWTAuth": []
                    },
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomObstacle"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/user/templates/{id}/create-design/": {
            "post": {
                "operationId": "user_templates_create_design_create",
                "description": "路线模板视图集。",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "标识此 路线模板 的 唯一整数值。",
                        "required": true
                    }
                ],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseTemplateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseTemplateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "CookieJWTAuth": []
                    },
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseTemplate"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/user/templates/{id}/favorite/": {
            "post": {
                "operationId": "user_templates_favorite_create",
                "description": "路线模板视图集。",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "标识此 路线模板 的 唯一整数值。",
                        "required": true
                    }
                ],
                "tags": [
                    "user"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseTemplateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseTemplateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "CookieJWTAuth": []
                    },
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseTemplate"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CourseTemplate": {
                "type": "object",
                "description": "路线模板序列化器。",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "title": {
                        "type": "string",
                        "title": "模板标题",
                        "maxLength": 120
                    },
                    "description": {
                        "type": "string",
                        "nullable": true,
                        "title": "模板描述"
                    },
                    "difficulty": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/DifficultyEnum"
                            }
                        ],
                        "title": "难度"
                    },
                    "field_width": {
                        "type": "integer",
                        "maximum": 4294967295,
                        "minimum": 0,
                        "format": "int64",
                        "title": "场地宽度"
                    },
                    "field_height": {
                        "type": "integer",
                        "maximum": 4294967295,
                        "minimum": 0,
                        "format": "int64",
                        "title": "场地高度"
                    },
                    "obstacle_count": {
                        "type": "integer",
                        "maximum": 4294967295,
                        "minimum": 0,
                        "format": "int64",
                        "title": "障碍数量"
                    },
                    "course_data": {
                        "title": "路线数据"
                    },
                    "cover_image": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true,
                        "title": "封面图"
                    },
                    "is_public": {
                        "type": "boolean",
                        "title": "是否公开"
                    },
                    "is_official": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "是否官方"
                    },
                    "author": {
                        "type": "integer",
                        "readOnly": true,
                        "title": "作者"
                    },
                    "author_username": {
                        "type": "string",
                        "nullable": true,
                        "readOnly": true
                    },
                    "copy_count": {
                        "type": "integer",
                        "readOnly": true,
                        "title": "复制次数"
                    },
                    "favorite_count": {
                        "type": "integer",
                        "readOnly": true,
                        "title": "收藏次数"
                    },
                    "is_favorited": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "创建时间"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "更新时间"
                    }
                },
                "required": [
                    "author",
                    "author_username",
                    "copy_count",
                    "created_at",
                    "favorite_count",
                    "id",
                    "is_favorited",
                    "is_official",
                    "title",
                    "updated_at"
                ]
            },
            "CourseTemplateRequest": {
                "type": "object",
                "description": "路线模板序列化器。",
                "properties": {
                    "title": {
                        "type": "string",
                        "minLength": 1,
                        "title": "模板标题",
                        "maxLength": 120
                    },
                    "description": {
                        "type": "string",
                        "nullable": true,
                        "title": "模板描述"
                    },
                    "difficulty": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/DifficultyEnum"
                            }
                        ],
                        "title": "难度"
                    },
                    "field_width": {
                        "type": "integer",
                        "maximum": 4294967295,
                        "minimum": 0,
                        "format": "int64",
                        "title": "场地宽度"
                    },
                    "field_height": {
                        "type": "integer",
                        "maximum": 4294967295,
                        "minimum": 0,
                        "format": "int64",
                        "title": "场地高度"
                    },
                    "obstacle_count": {
                        "type": "integer",
                        "maximum": 4294967295,
                        "minimum": 0,
                        "format": "int64",
                        "title": "障碍数量"
                    },
                    "course_data": {
                        "title": "路线数据"
                    },
                    "cover_image": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true,
                        "title": "封面图"
                    },
                    "is_public": {
                        "type": "boolean",
                        "title": "是否公开"
                    }
                },
                "required": [
                    "title"
                ]
            },
...
            "CustomObstacle": {
                "type": "object",
                "description": "自定义障碍物序列化器",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "title": "障碍物名称",
                        "maxLength": 100
                    },
                    "obstacle_data": {
                        "title": "障碍物数据"
                    },
                    "user": {
                        "type": "integer",
                        "readOnly": true,
                        "title": "用户"
                    },
                    "user_username": {
                        "type": "string",
                        "description": "获取用户名",
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "创建时间"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "更新时间"
                    },
                    "is_shared": {
                        "type": "boolean",
                        "title": "是否共享"
                    }
                },
                "required": [
                    "created_at",
                    "id",
                    "name",
                    "obstacle_data",
                    "updated_at",
                    "user",
                    "user_username"
                ]
            },
            "CustomObstacleRequest": {
                "type": "object",
                "description": "自定义障碍物序列化器",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "title": "障碍物名称",
                        "maxLength": 100
                    },
                    "obstacle_data": {
                        "title": "障碍物数据"
                    },
                    "is_shared": {
                        "type": "boolean",
                        "title": "是否共享"
                    }
                },
                "required": [
                    "name",
                    "obstacle_data"
                ]
            },
...
            "DifficultyEnum": {
                "enum": [
                    "easy",
                    "medium",
                    "hard"
                ],
                "type": "string",
                "description": "* `easy` - 初级\n* `medium` - 中级\n* `hard` - 高级"
            },
...
        },
        "securitySchemes": {
            "CookieJWTAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token",
                "description": "登录后由后端写入的 httpOnly JWT Cookie"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}